requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0
pytest>=7.4.0
pytest-mock>=3.11.0
responses>=0.23.0